Add post-install hook to create a conda environment.
When creating a project a conda environment is also now created, using
the python version specified. The envrionment is also pre-installed
with some simple dev tools (pytest, flake8 and invoke).

Note that tox and conda don't always play nicely together, and as such
tox is not available within conda and hence not installed in the
environment. It can be installed with `pip install tox` after
activating the environment, and also requires having the conda version
of `virtualenv` installed in the root conda environment.

For more information on this see `https://bitbucket.org/hpk42/tox/issues/273/support-conda-envs-when-using-miniconda`